requirements.txt
django
djangorestframework
mysqlclient
apkutils